Rename: 영상게시판 스타일 파일 이름 변경
